refactor: 음식점 상세 정보 모달 관련 폴더 구조 수정
