+++ client/src/components/RestaurantDetailModal/RestaurantDetailModalCarousel/styles.tsx
